moving const, path and conf file to a single dir
