+++ get_data/src/get_from_db.py
@@ -4,7 +4,7 @@ from tqdm import tqdm
 
 from get_data.src import es_utils
 from get_data.src import s3_utils
-from get_data.cluster_param import ES_INDEX, ES_HOST_IP, ES_HOST_PORT
+from conf.cluster_param import ES_INDEX, ES_HOST_IP, ES_HOST_PORT
 
 
 def get_missing_picture(picture_dir, l_wanted_pic):
