+++ get_data/src/label_handler.py
@@ -2,8 +2,8 @@ import json
 from pathlib import Path
 
 from utils import car_mapping as cm
-from utils.path import SESSION_TEMPLATE_NAME, HARDWARE_CONF_FILE
-from utils.const import SPEED_FAST, SPEED_NORMAL, IMAGE_SIZE, FRAME_RATE, EXPOSURE_MODE, \
+from conf.path import SESSION_TEMPLATE_NAME, HARDWARE_CONF_FILE
+from conf.const import SPEED_FAST, SPEED_NORMAL, IMAGE_SIZE, FRAME_RATE, EXPOSURE_MODE, \
     DIRECTION_C, DIRECTION_L, DIRECTION_L_M, DIRECTION_R, DIRECTION_R_M, HEAD_DOWN, HEAD_UP, \
     MAX_SPEED, STOP_SPEED, MAX_DIRECTION_LEFT, MAX_DIRECTION_RIGHT
