+++ get_data/src/update_db.py
@@ -1,7 +1,7 @@
 
 from get_data.src import s3_utils
 from get_data.src import es_utils
-from get_data.cluster_param import ES_INDEX, ES_HOST_PORT, ES_HOST_IP
+from conf.cluster_param import ES_INDEX, ES_HOST_PORT, ES_HOST_IP
 
 
 def delete_picture(picture_id, bucket_path):
