+++ utils/car_mapping.py
@@ -1,5 +1,5 @@
-from utils.const import SPEED_NORMAL, SPEED_FAST, DIRECTION_L_M, DIRECTION_C, DIRECTION_L, DIRECTION_R, DIRECTION_R_M
-from utils.const import MAX_DIRECTION_LEFT, MAX_DIRECTION_RIGHT, MAX_SPEED, STOP_SPEED
+from conf.const import SPEED_NORMAL, SPEED_FAST, DIRECTION_L_M, DIRECTION_C, DIRECTION_L, DIRECTION_R, DIRECTION_R_M
+from conf.const import MAX_DIRECTION_LEFT, MAX_DIRECTION_RIGHT, MAX_SPEED, STOP_SPEED
 
 
 def _two_label_speed_mapping(speed):
